requirements.txt
backports.shutil-get-terminal-size==1.0.0
certifi==2017.4.17
chardet==3.0.4
decorator==4.1.2
Django==1.11.3
django-formtools==2.0
django-pipeline==1.6.13
django-polymorphic==1.2
djangorestframework==3.6.3
docx==0.2.4
enum34==1.1.6
futures==3.1.1
html5lib==0.9999999
idna==2.5
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
lxml==3.8.0
messagebird==1.2.0
MySQL-python==1.2.5
olefile==0.44
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
pkg-resources==0.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.2
pycrypto==2.6.1
Pygments==2.2.0
pyPdf==1.13
PyPDF2==1.26.0
python-dateutil==2.6.1
python-magic==0.4.13
pytz==2017.2
reportlab==3.4.0
requests==2.18.1
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
South==1.0.2
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
webencodings==0.5.1
xhtml2pdf==0.0.6
